I need to sleep
There's some debug stuff now, and a way to completely bypass the parser.
That's cool I guess. Also less game-breaking exceptions when doing random
things. debug is very fragile. debug is dry clean only. Do not get the
debug in your eyes. If you do, contact poison control immediately. debug
is not to be used by women who are pregnant or may become pregnant. If you
only use debug for greater than 48 hours, please contact a core coder for
assistance and a 12-step program. If the cost of debug is too much for
you, BoofPC may be able to help.

PAID ACTORS