YOU CAN NOW WIN THE GAME!!
There are still a few minor problems that will need fixed- for example,
in order to leave the staff lounge you have to have the lantern. You
should be able to leave even without the lantern. We still need to add
points in a couple places, other than just the couple of easter eggs.
Oh- and Hallway 14 still needs a description. Overall, it is coming
together quite nicely. Good job everyone!